Bootloader exceeds 512 byte limit.
The bootloader is now larger than 512 bytes, allthough
the code within the first 512 bytes is still self contained,
with one reference to a function call starting at offset 512.

It now uses genisoimage -no-emul-boot -boot-load-size 2,
instructing the BIOS to load multiple sectors, as genisoimage
refused to write this information in the ISO9660 image.