Initial memory management.
Uses sbrk() system call to extend the end of the data segment.
Keeps a list of handles of memory given to malloc() callers
for use in free(). Keeps an ordered list of free()d segments,
low to high address, in order to more easily merge adjoining
segments to remove handles.

The handle list is allocated using internal_malloc. Since it is
private, it can be reallocated.

The main function operates a bit like the unix ''free'' utility.
The underlying management can also be used for global memory management;
the sbrk() will then be unnecessary.

Uses CPP for allowing declaration of variables in mid-method.
Uses the stdc++ lib to allow struct foo = { ... }

Contains a bug in ordering the free segment linked list.