Fix calculation of load sectors. Adjust load addr.
Using kludge to calculate the exact number of sectors to write.
Adjust the start offset stored in the initial registers on
stack for re-use in writebootsectors.

Writebootsector disabled, but runs succesfully; however,
the system does not recognize the bootsector when booting
from primary harddisk.

Minor adjustments to preserve bp, removed some debug.