Relocate bootloader/kernel. Fix pmode. Add ACPI.
Added code to relocate the bootloader to 0x100000 (seg 1).
This created a bug in the pmode using wrong offsets. Fixed.

Menu: add list floppies, using io ports.

ACPI: code is stable yet not functioning. The reference to the ACPI
datastructures needed appears to be located at 07ff0000, beyond
the reach of real mode.

Renamed print32.s to print2.s, for additional printing functions
that would increase the size of the bootsector too much.

Added util/checksymbols to check for missing symbols in the bootloader.o
file. As it is not linked, this is not checked, and these references
get silently replaced by 0.

Minor work in inspectmem: add a 'r' to reset segment/offset to 0,
and have the left/right arrow keys change the segment value,
so that the entire real mode address space can be inspected.