Bootloader: Fix restore signature; fix call addr.
The signature restored at the end of the bootsector in memory,
which is used for the initial stack, was restored incorrectly
(endianness).

Further, the return address of the 1f call in black, used to
correct the starting offset of the bootloader code, was off.

Also a minor fix where the print.s was included within the .if WHITE
code, yet, this code needs to be available (printhex) for
the black loader.