Pmode: add SS,ES,FS,GS selectors: share stack.
The Realmode SS segment is converted to a SS selector, thus
allowing to share the stack between realmode and pmode, or
to at least access the realmode stack without modifying the ESP
register (except to zero the top word which isn't used in realmode
stack operations).

The ES segment is also converted to a selector, which at current
demonstrates textmode video buffer access where EDI is zero-based,
making calculations easier.

Two selectors based in 0xb8000 and 0xa00000 are added to provide
this ease-of-access.

The other segments FS, GS are also converted to selectors
to create realmode-pmode interoperabilty, potentially allowing
the realmode code to run in pmode using the same offsets.