PMode: Allow enter/exit separately; add TSS.
Before this change:

	.code16
	call  protected_mode

	protected_mode:  // called from realmode
	  enter pmode
	  .code32
	  payload
	  enter realmode
	  .code16
	  return to realmode caller

This is now:

	.code16
	call  protected_mode
	call  real_mode

	protected_mode: // called from realmode
	  enter pmode
	  .code32
	  payload
	  convert return address to 32 bit
	  return to pmode return address on stack

	real_mode: // called from pmode
	  enter realmode
	  .code16
	  convert return address to 16 bit
	  return to realmode return address

The protected_mode function now returns to it's real-mode caller
in protected mode which then has to call the real_mode function itself.
The function translates the 16 bit return address on the stack
to the proper 32 bit address.

The code segment used is a 32 bit version of the 16 bit realmode CS,
so that offsets are identical, and 16 and 32 bit code can be mixed
in one segment.

The ds, es and ss selectors are 16 bit by default, so that only
the code is now 32 bit accessible.

The real_mode function now assumes that it was called from protected
mode, whereas before this code would have been jumped to from
pmode, leaving the stack value return address to 16 bits.