Bootloader: calculate proper stack.
As sector1 is growing, it is now entering the 0x2000 stackspace. So,
code_size is added and stack is quaranteed to be 0x1f00 to 0x2000 in size.
Since this stack is only used in sector1 which job it is to load the
kernel (right now it is a monolithic kernel), and go into pmode,
a new stack will be allocated soon. Currently the pmode places the
Task State Segment stack for Interrupt Gates at the start of the binary,
as to provide a known location regardless of where it is loaded/relocated.

Also, restore the original sp before it is set to 512 - to enable the call
to determine the loaded address to be known - into the new stack's saved
sp (which would contain the preprogrammed value of 512). Thus now
the entire original register set is preserved, except for the flags.