Major overhaul: split kernel from loader.
The bootloader still loads sector1, however sector 1 is now
dedicated to loading whatever sector follows it on disk, and see if it
is a ramdisk. If so, it will load the first entry and transfer control to it.

The second image is a separate kernel. At current it reproduces a few
16 bit print functions, used for debugging the transfer to protected mode.

Defunct code: most of what is in 16/ is now not used. It can be easily
embedded within sector1 by including '16/menu'. In this case,
the code in sector1 will need to go into a menu item.