Fix GPF in std int handler & cls. Add PCI lister.
The clear screen function used a selector constant defined in gdt.s which
wasnt included. The assembler and the symbolchecker failed to report this,
causing ds to be loaded with a memory reference generating a GPF.

The exceptionhandler itself made a similar mistake by trusting the value
on the stack as a valid selector, causing recursive faults. Checks are
now in place.

Also, some exception interrupts push error values on the stack. Stack
analysis is adjusted.

An initial PCI device lister is added, though it appears to be nonfunctional.