Exception handler improvement.
Printing the proper exception name now works. Also, a check is added
to see whether the stack conforms to a software INT, which does not
push an error code onto the stack for certain exceptions. This check
might yield a false positive for an INT invocation when EIP happens
to fit selector criteria and an error code is present on the stack.
The CPL/RPL/DPL is filtered out properly from the examined selector.
The default behaviour of an Exception (INT < 0x20) is to halt the system.

Printing update: some minor fixes and improvements. Tried using
the .altmacro and LOCAL for labels, but this did not work well
as string interpretation (using '', ! and such) is altered.
print.s now uses label 99 for strings and jumps within macro's.

Update dt: SEL_MAX now includes RPL 3 bits for ease of checking in the
exception ISR.