PMode improvements, separate stack.
Rework protected/realmode switching code. When calling protected_mode
with ax = 1, a separate pmode stack is allocated, using the ds selector,
so that the bases are aligned, and things like 'rep movsb' work between
ds and ss.

The code selector limit in gdt.s is no longer based on the kernel code size.

The [realsegflat] is reset to 0 when the code base aligns with realmode cs.

The re-enter protected mode code will use the same settings as when
pmode was first initialized.

The call_realmode code is rewritten. Due to the version of GAS not
supporting .pushsection or .previous, using print macros in .text32
is unstable as it switches to the .text16 segment after allocating the
string. Thus the code switch between realmode/pmode and vice versa
does not use contiguous code, but uses .text 0 and .text 3 sections.

Lots of debugging/printing cleanups.

Shell adds 'gpf' command to trigger a general protection fault.
Fixed stack ref bug in idt.s exception/interrupt handler.