Filesystem improvements.
Rewrote file/directory opening using a string tokenizer,
and using the stack to allocate the fs_handle structure used during
directory traversal. Upon success it will be copied to a newly
allocated fs_handle. This cleans up the use of many registers for
handles.

Mounting on top of existing mounts (like the '/ (none)') works,
by using the last matching mountpoint.

The use of the stack, and passing registers referencing it required
to have the stack base equal to the data base, which is implemented
in the previous commit.

string.s:
	add strlen_ using esi->ecx (instead of eax->eax for strlen).
	add strtok for a single separator char.

fat.s: fix cluster to sector conversion for file reading.