Update TODO
made some notes in the source and TODO regarding I/O pipes,
multitasking, keyboard, string data optimization.

Added a .strtab section in the .o file, which gets removed
from the .bin file (has to be an explicit parameter to objcopy,
otherwise it still doesn't copy the section, but it doesn't do
proper relocation - the .bin is somewhat garbled).