Net queue scheduled once.
The netq task is scheduled only once, and keeps processing the
queue, yielding when it is empty.

TODO: implement IO scheduling mutexes, where the scheduler does not
schedule a process unless some mutex is set.
It should also be possible to schedule the process regardless periodically
so it can do tcp connection cleanup.
Possible implementation is to set a timeout for rescheduling, which will
be reset by the net_rx_packet to indicate the netq should be scheduled ASAP.