Refactor filesystem to OO.
Some debug present.

Also contains a hack for fs_api_mount, which 'instantiates' an object
so that the method pointers are filled in, as it is actually a factory
method.

Contains work on static methods, as yet unused.

Refactored the class definition to use more data subsections, where
the declarations, overrides, and static data are all separated and
have their own pointers.
Unfortunately GAS by itself does not have enough flexibility to allow
for custom data structures to be pipelined. The class definition
could be made more compact, first by declaring all method pointers
in the class definition itself as it allows for this by using a
size field.
The upside of the current approach is that method declarations of different
types (declare, override, static) can be mixed.

Note that the work done on static class methods is not quite how it is
commonly defined in OO. In this case the meaning of static is that
the method is given the class definitino itself as the object.
Also, the static method is envisioned to also be allowed to be overridden.
As such it is a copy of the normal declaration/override system, except
the object it applies to is the class's static data.

The intended use for this is for factory methods, such as fs_api_mount,
which contain a few checks to see whether the partition type matches the
filesystem implementation, and if so, it will instantiate and mount
the partition.
In this way, mount expects all classes extending from fs to support
this static method.

This commit however uses a temporary object instantiation, calling
fs_api_mount in ojbect context, which then proceeds to class_newinstance
an object, initializing it, and returning it. The temporary object is
then freed. This workaround was done so as to gradually change the
implementation of the various filesystems, which used to do a malloc
themselves, but now call class_newinstance.

There is rather a lot of debug output in this commit, but, it is the
first attempt at using git branches, so more commits are expected
before a merge.