httpd: use small buffer for files
Instead of loading the entire file at once, read and send the
content in chunks (at current 2kb, the ISO sector size; tested
200kb which makes no difference in speed).
Due to this, the ${www-expr} functionality has _temporarily_
been disabled since these can cross buffer boundaries.

This commit fixes the pagefault when the iso is downloaded due to
the buffer requirements exceeding the paged memory range.

Also, the download speed is at current limited to about 350kb/s
with the PIT set to 250Hz. Tests show that setting the PIT to 500Hz
doubles the network speed, and setting it to 1024 or higher yields
a top network speed of over 5.3Mb/sec.
It appears that perhaps the idle task gets scheduled resulting in
a wait for the next interrupt.