implement 16-bit kernel relocation
After doing an upgrade to GCC 4.8.2 the kernel no longer
booted. This commit is a step in the direction of having
the kernel boot again, by implementing 16bit relocation.

As it turns out, the new GCC objcopy or ld no longer resolves
16bit data references while folding data into code.

Two main changes are implemented in the relocation table code:

1) the 16bit relocation table is written. A bug is now fixed
that would make the 32 bit relocation table seem empty when 16
bit relocation data was written. Now, there is always a count
for 16 bit relocation entries regardless of whether 16 bit relocation
is clobbered, thus having the bootloader properly detect the correct
offset for the 32 bit relocation table.

2) there are now 2 16 bit tables, one for .data16 and one for .text,
as each has a different relocation.
Since it is required that the 16 bit kernel header is executed in
realmode, the kernel must not be loaded high. Therefore, the 32 bit
relocation must be within 16 bit address space so as to not overflow
the 16 bit relocation table.

Makefile:
	temporarily produce a minimal ISO restoring the
	initial build speed. The idea is to produce a second
	DATA ISO which' changes are mostly independent of the
	code or Kernel ISO containing the boot image and a minimal
	filesystem (with the default minimalistic site sans documentation
	and source references).
	Basically all generated and imported web content is no longer
	imported.

kernel/realmode.s:
	add debugging to test realmode data offsets. This revealed that
	the .data16 offsets were indeed relative to .data16 ( > 0)
	and no longer normalized to .text.

bootloader/sector1.s:
	moved the pmode code into a 16 bit shareable include
	in preparation of reusing the code in kernel/pmode.s,
	as it contains a working interrupt handler. The pmode.s
	does not set up the IDT until protected mode is entered,
	and so this minimal handler would assist in debugging
	the current boot problem described in pmode.s's changelog.

16/pmode.s:
	factored out of bootloader/sector1.s.

kernel/gdt.s:
	debugging update; print the cs selector data.

kernel/pmode.s:
	QEmu's debug logging reveals that a protection fault
	is generated in the 'jlmpl' instruction right after
	setting the PMODE flag in cr0.
	The addressing scheme has been modified in setting up
	the protected mode address to jump to. It is as yet
	uncertain which is the right value, although the current
	value seems correct. 16 bit symbol addressing is now used,
	so that the 16 bit relocation is applied to the pmode_entry$
	method aswell. (otherwise the address ends up being relocated
	not at all or only with the .text offset (0x13000), and thus
	missing the offset of .data16 which _should_ be 0 but is
	relative to .text/.text16).