[WIP] booting proceeds until hardware int
Further improvements; this commit marks a stable situation.

The screen_buf_flush event handler in kernel/print.s is disabled
due to false symbol relocation.

Updated the GDT and IDT code to use 16 bit offsets to fix relocation
for the .data16 section (same as before; when referring to data in
.data16 there are 2 relocations: 32bit .text (0x13000), and 16bit .data16
(0x2000 or so).