util/adm vm: fixes; allow name for node
temporarily added 'use strict' to test warnings, but the
code makes use of non-strict refs and subs (which I failed
to be able to disable).

Argument passing and managed node name testing now accepts
a name aswell as a number: /\d+|\w+/;

This is so that a dedicated node can be spawned, in this case,
one for testing the latest features, keeping the other nodes stable;
they will then all be named for their function if needed.
(I now use the numbers: 0 = DMZ, 1 = fallback, etc).