cloudnet: uphold DMZ IP service
The DMZ IP is configured in 2 places:
- nic.s (nic_zeroconf)
- cloudnet.s (DMZ_IP)

cloudnet will attempt (gratuitious ARP) to use the IP the nic is configured
with, which is the zeroconf IP, and fallback to DHCP.

Every minute, after having gathered node pings, it will check whether
the DMZ IP cluster node is up. It will then obtain the DMZ IP
using gratuitious ARP.
(note: this attempt will also be made when the IP is taken but does
not belong to a cluster node).

See the TODO in DOC in this commit.

Also TODO: configure the DMZ IP in the VMX file and in DNS.
.cloud is in the process of becoming a TLD (or has already)
so the bootstrap domain will be used with a virtual view on
DNS - i.e., within a LAN, a lookup for 'dmz.cloud.neonics.com'
would resolve the IP to use, whereas it is invisible from the
public IP space.