ps: show registered CPL
Since all tasks are in blocking IO mode most of the time,
the displayed CPL level (the P column for Privilege level)
usually displays '0', the scheduler's CPL.

This commit compresses the 'flags' column (the " T" (task) and "J " (job)
fields are now overlapping since it is either/or) and appends the
CPL level from the TASK_FLAGS. This _should_ match the DPL of cs
in the userspace stack.

A following commit will extend the 'ps' command with a 'pid' parameter
displaying detailed task information.