Fix concurrency-NPE, data-loss in sample code
The old code:

	alert_init:
		mov	eax, CONST
		xchg	eax, [A]
		mov	[B], eax

suffers from a potential concurrency issue between the last two lines;
an interrupt might prompt a screen update, handled by the alert code,
which delegates to [B], which is as yet not initialized.

In terms of efficiency, the old code contains 3 memory references (two
in the xchg (read and write), and 1 in the final statement),
and 4 register references (constants in the opcode can be described by
a 'const' register referencing the area in the opcode containing the value).

The new code:
	mov	eax, [A]
	mov	[B], eax
	mov	[A], CONST

contains 3 memory references, and 3 register references, and might thus be faster.

The worst solution to fix the concurrency issue would be to conditionally call
the proxied handler, the original [A], as this incurs:

	cmpd	[B], 0
	jz	$skip

on each iteration. (Calls to 'alert_init' are also considered iterations, but
of a lower octave, a greater scope: task- / process- / thread- lifecycles).

A check has also been added to ensure that the original value of [A],
preserved in [B], cannot be lost by constant reduction. Constant
reduction is the removal of a variable by replacing it with a constant,
an operation that loses entropy, so to speak. In this case, a second
invocation of alert_init would have [A] = [B] = CONST, which looses the
ability to successfully wrap the original [A] by not preserving it. A
package that looses it's content is hardly fulfilling it's function to
hold it's content.