netq: fix bufferoverflow on rx large packets
Since the eth frame size has been increased to the official maximum, some
incoming responses packets now also exceed 1500 bytes, corrupting the malloc
pool if such a the packet happens to be copied to the last packet buffer.
This commit adds a constant for the proper maximum size and updates the
large packet check to enforce exactly this size.

kernel/net/net.s:
	rename BUFFER_SIZE to BUFFER_TX_SIZE;
	replace 1500 with new BUFFER_RX_SIZE, size 0x600 (1536);
	reduce checked size of large packet check from 2000 to
	BUFFER_RX_SIZE;
	remove redundant [net_rx_queue] null check (the mov eax
	is also redundant but there is still some uncommitted
	debug above it, trashing eax).